Separate mainFunction in other file
