+++ index.js
@@ -1,5 +1,9 @@
 /* eslint-disable indent */
 const Scrapper = require('./webScraping/cfeScrapper');
+const Validation = require('./utils/mainFunction');
 const text = 'IoT';
+const route = './assets/Data-From-Table.json';
 const myScrap = new Scrapper(text);
-myScrap.mainFunction();
+const myValidator = new Validation(route);
+
+myValidator.mainFunction();
