+++ webScraping/cfeScrapper.js
@@ -10,11 +10,9 @@
  */
 
 const puppeteer = require('puppeteer');
-const fs = require('fs');
 const ScrapPage = require('../scrapperFunctions/index');
 const URLPage = 'https://msc.cfe.mx/Aplicaciones/NCFE/Concursos/';
 const idInput = '#descProc';
-
 const idButton = '#buscar';
 const waitingTime = 2000;
 const route = './assets/Data-From-Table.json';
@@ -119,33 +117,6 @@ Scrapper.prototype.doScraping = async function () {
     console.error(`Error: ${err}`);
     throw err;
   }
-};
-
-/**
- * Compares the date´s file and today to check if doScraping or not
- */
-// eslint-disable-next-line space-before-function-paren
-Scrapper.prototype.mainFunction = function () {
-  // eslint-disable-next-line space-before-function-paren
-  fs.stat(route, (err, stats) => {
-    if (err) {
-      this.doScraping();
-    } else {
-      let dateLastModified = stats.mtime;
-      let dateToday = new Date();
-
-      dateLastModified = dateLastModified.getTime();
-      dateToday = dateToday.getTime();
-
-      const dif = (dateToday - dateLastModified) / (1000 * 60 * 60);
-      if (dif > 20) {
-        this.doScraping();
-      } else {
-        console.log('Scrap completed correctly');
-        console.log(`The data has been saved in: ${route}`);
-      }
-    }
-  });
 };
 
 module.exports = Scrapper;
